Step 5: Split pom/model.py into line_roles.py and dependencies.py
Extracts SectionRole, ROLE_FACTORS, build_line_roles into line_roles.py
and PomDependency into dependencies.py. Decomposes _build_model into
_extract_coordinates, _extract_parent, _extract_dependencies, _line_of.
