+++ src/oss_sanitizer/scanners/pom/__init__.py
@@ -1,13 +1,8 @@
 """POM model sub-package."""
 
-from .model import (
-    parse,
-    parse_from_text,
-    PomModel,
-    PomDependency,
-    SectionRole,
-    ROLE_FACTORS,
-)
+from .dependencies import PomDependency
+from .line_roles import ROLE_FACTORS, SectionRole
+from .model import PomModel, parse, parse_from_text
 
 __all__ = [
     "parse",
